### 2020-05-22 #2224 @whedon
[PRE REVIEW]: multivar_horner: A Python package for computing Horner factorisations of multivariate polynomials
**Submitting author:** @MrMinimal64 (<a href="http://orcid.org/0000-0002-1819-6975">Jannik Michelfeit</a>)
**Repository:** <a href="https://github.com/MrMinimal64/multivar_horner" target ="_blank">https://github.com/MrMinimal64/multivar_horner</a>
**Version:** v.2.0.0
