### 2020-05-23 #2224 @arfon
:wave: @MrMinimal64  - thanks for your submission to JOSS. Could you please add a few sentences to your paper to help those less intensely involved in mathematics understand the downstream applications of this work?  That would help to assess the impact of the effort and help clarify the range of research users who would be impacted by this.
