### 2020-05-23 #2224 @MrMinimal64
Hello @arfon, thanks for your quick response. Could you tell me which parts I should clarify or which parts are unclear? I thought I had described the advantages and potential use cases of the package extensively in the first section "**Summary**". The following sections give a more detailed explanation of the implemented approach and underpin the advantages with benchmarks.

I am looking forward to your review.
