### 2020-05-24 #2224 @arfon
> Hello @arfon, thanks for your quick response. Could you tell me which parts I should clarify or which parts are unclear?

I think a little more detail in the summary section would be useful. For example, are there applied research fields where the factorisation of multivariate polynomials is a key aspect of the research work/analysis?

> should I suggest reviewers myself?

Yes, please go ahead but as I [mentioned above](https://github.com/openjournals/joss-reviews/issues/2224#issuecomment-633123110), we won't actually begin the review process until an editor is assigned to this paper which may be some time due to a large backlog of submissions.
