### 2020-05-25 fd48f53
work on paper
clarify possible applications in paper
work on documentation
