+++ paper/paper.bib
@@ -66,6 +66,12 @@ series = {LLVM ’15}
   publisher={IEEE}
 }
 
+@article{michelfeitresponse,
+  title={A Response Surface Method based on Multivariate Newton Interpolation},
+  author={Michelfeit, Jannik}
+  year={2019}
+  url = {https://mosaic.mpi-cbg.de/docs/Michelfeit2019.pdf},
+}
 
 @article{Hecht1,
   title={Multivariate {N}ewton Interpolation},
