+++ requirements_tests.in
@@ -14,5 +14,5 @@ flake8-commas
 isort # to show where imports aren’t sorted
 multilint
 pygments # to check your setup.py is well configured
-
+rstcheck>=3.3.1
 
