### 2020-05-25 #2224 @MrMinimal64
I clarified the possible application of this package in the paper and added example applications where the package is already being used.
