### 2020-05-25 #2224 @MrMinimal64
Based on the topics (applied mathematics, numerical analysis...) and preferred language (Python) they provided, I suggest the following reviewers:

```
munkm
SebastianoF
diehlpk
crstnbr
mstimberg
cpgr
fccoelho
bonh
dglmoore
```
