+++ paper/paper.bib
@@ -66,12 +66,18 @@ series = {LLVM ’15}
   publisher={IEEE}
 }
 
+@article{michelfeitresponse,
+  title={A Response Surface Method based on Multivariate Newton Interpolation},
+  author={Michelfeit, Jannik},
+  year={2019},
+  url = {https://mosaic.mpi-cbg.de/docs/Michelfeit2019.pdf},
+}
 
 @article{Hecht1,
   title={Multivariate {N}ewton Interpolation},
   author={Hecht, Michael and Hoffmann, Karl B and Cheeseman, Bevan L and Sbalzarini, Ivo F},
   journal={arXiv preprint arXiv:1812.04256},
-  year={2018}
+  year={2018},
 }
 
 
