+++ paper/paper.bib
@@ -4,7 +4,7 @@
   title = {multivar_horner},
   year = 2018,
   url = {https://github.com/MrMinimal64/multivar_horner},
-  urldate = {2020-04-20}
+  urldate = {2020-05-28},
 }
 
 @article{numpy,
@@ -71,6 +71,7 @@ series = {LLVM ’15}
   author={Michelfeit, Jannik},
   year={2019},
   url = {https://mosaic.mpi-cbg.de/docs/Michelfeit2019.pdf},
+  urldate = {2020-05-28},
 }
 
 @article{Hecht1,
