### 2020-05-29 cd211c0
mention BUGFIX in Changelog
