### 2020-05-29 #2224 @dpsanders
👋 Hi @MrMinimal64. Thanks for submitting to JOSS!

I have been assigned as the editor for your submission. Looking over it, I have a few questions that I would like you to address in the paper / docs. They are a mixture of technical and about the impact of the submission.

- There are several representations of multivariate polynomials. Does your package specify or require one? Are you basically providing an implementation of multivariate polynomial manipulation? Does it only work with that representation?

- Are there related Python packages for manipulating multivariate polynomials? Can your package interoperate with them?

- Can you define more precisely what you mean by "canonical form" and “Horner factorisation”, and the advantages that the latter has?

- It would be helpful if you gave an example of a 1D Horner factorisation as an example / motivation.

- You should show the actual output that your code gives when you calculate the Horner factorisation.

- How is the usefulness of this impacted by Python's speed? Do you automatically use Numba?

- Please add points on figure 1. Some of the figures do not need to be so large.

- How do you calculate the numerical error in your evaluations?

- What are the units for the error in figure 3?

- It seems that the main application may be to multivariate interpolation. Do you have an implementation of this?

- The JOSS editors requested that you add a paragraph to help a more general audience understand the impact and applications of your work, and to clarify the range of research users who would be impacted.

Thanks!
