### 2020-05-29 #2224 @dpsanders
Also, a comparison with the Julia package https://github.com/JuliaAlgebra/StaticPolynomials.jl may be interesting, since that seems very relevant / related.
