### 2020-06-03 6a8e817
incorporate editor feedback
