+++ paper/paper.bib
@@ -66,6 +66,14 @@ series = {LLVM ’15}
   publisher={IEEE}
 }
 
+@book{neumaier2001introduction,
+  title={Introduction to numerical analysis},
+  author={Neumaier, Arnold},
+  year={2001},
+  publisher={Cambridge University Press}
+}
+
+
 @article{michelfeitresponse,
   title={A Response Surface Method based on Multivariate Newton Interpolation},
   author={Michelfeit, Jannik},
