+++ paper/paper.md
@@ -29,14 +29,17 @@ bibliography: paper.bib
 # Summary
 
 Polynomials are a central concept in mathematics and find application in a wide range of fields.
-Thereby, one often asks for numerically stable and efficient evaluation schemes, which mostly rely on compact representations of (multivariate) polynomials [@LeeFactorization2013; @leiserson2010efficient; @Hecht1].
+(Multivariate) polynomials have different possible mathematical representations and the beneficial properties of some representations are in great demand in many applications[@LeeFactorization2013; @leiserson2010efficient; @Hecht1].
 
-The `multivar_horner` Python package implements a multivariate Horner scheme ("Horner's method", "Horner's rule") [@horner1819xxi] and thereby allows computing Horner factorisations of multivariate polynomials.
-Compared to the canonical form of polynomials this representation offers some important advantages.
-The Horner factorisation is more compact in the sense that it requires less mathematical operations in order to evaluate the polynomial (cf. \autoref{fig:num_ops_growth}).
+
+The so called Horner factorisation is such a representation with beneficial properties.
+Compared to the canonical (unfactorised) form, this representation offers some important advantages.
+First of all the Horner factorisation is more compact in the sense that it requires less mathematical operations in order to evaluate the polynomial (cf. \autoref{fig:num_ops_growth}).
 Consequently, evaluating a multivariate polynomial in Horner factorisation is faster and numerically more stable [@pena2000multivariate; @pena2000multivariate2; @greedyHorner] (cf. \autoref{fig:num_err_growth}).
 These advantages come at the cost of an initial computational effort required to find the factorisation.
 
+The `multivar_horner` Python package implements a multivariate Horner scheme ("Horner's method", "Horner's rule") [@horner1819xxi] and thereby allows computing Horner factorisations of multivariate polynomials given in canonical form.
+Representing multivariate polynomials or arbitrary degree also in canonical form, computing derivatives of polynomials and evaluating polynomials at a given point are further features of the package.
 Accordingly the package presented here can be helpful always when (multivariate) polynomials have to be evaluated efficiently, the numerical error of the polynomial evaluation has to be small or a compact representation of the polynomial is required.
 This holds true for many applications applying numerical analysis.
 One example use case where this package is already being employed are novel response surface methods [@michelfeitresponse] based on multivariate Netwon interploation [@Hecht1].
@@ -44,32 +47,38 @@ One example use case where this package is already being employed are novel resp
 
 # Functionality
 
-In one dimension there is only a single possible Horner factorisation of a polynomial.
-In the multivariate case however the factorisation is ambiguous as there are multiple possible factors to factorise with.
-The key functionality of `multivar_horner` is finding a good instance among the many possible Horner factorisations of a multivariate polynomial.
+In its core `multivar_horner` implements a multivariate Horner scheme with the greedy heuristic presented in [@greedyHorner].
+In the following the key functionality of this package is being outlined.
+For a more details on polynomials and Horner factorisations please refer to the literature, e.g. [@neumaier2001introduction].
 
-Let's consider the example multivariate polynomial $p(x) = 5 + 1 x_1^3 x_2^1 + 2 x_1^2 x_3^1 + 3 x_1^1 x_2^1 x_3^1$.
[…]
