### 2020-06-03 #2224 @MrMinimal64
Hello @dpsanders,

Thank you for your detailed review.
I tried to address all of your open questions and remarks in the latest commit.
Please have a look at the new version of the paper.
Some points I could however not resolve:

> There are several representations of multivariate polynomials. Does your package specify or require one?

The core is “specifying” a Horner factorisation representation, but this “requires” the input of a canonical representation. Is this evident from the paper?

> Are you basically providing an implementation of multivariate polynomial manipulation?

In my understanding polynomial manipulation is about computations with polynomials, like addition etc.. This is not supported by `multivar_horner`. Consequently it is also not being mentioned in the paper or documentation. Should I specifically state that this is not within the scope of the package?

> Can you define more precisely […] the advantages that the [Horner factorisation] has?

I stated in the paper that the Horner factorisation has a lower numerical error and requires less operations during an evaluation. The included figures try to prove these claims. Do you have suggestions on how to define the advantages more precisely?

> It seems that the main application may be to multivariate interpolation. […] The JOSS editors requested that you add a paragraph […] to clarify the range of research users who would be impacted.

The functionality of this package is computing an advantageous representation of multivariate polynomials. Since this contribution is so general, it is hard, and would in my opinion even be misleading, to give specific use cases (apart from examples).

> The JOSS editors requested that you add a paragraph to help a more general audience understand the impact and applications of your work […].

What aspects of the approach remain unclear (for a more general audience)? The way I see it, an explanation for a wider audience would require the definition of very basic mathematical concepts like polynomials etc.. For the intended audience (= potential users of the presented package) some basic knowledge about these concepts can, and i even think must, be assumed. Everything else would blow up the scope of the paper and also obscure the relevant information for the intended audience. Would it not? Please correct me if I am routine-blinded and some aspects of the presented approach are incomprehensible even to readers with prior knowledge.

> Some of the figures do not need to be so large

How do I adjust the width of the figures? It seems to me that the markdown standard itself does not specify this.

> What are the units for the error in figure 3?

Since the given values are relative (to the numerical error of evaluation with Horner factorisation), there is no unit.

> Do you have an implementation of this [multivariate interpolation]?

Yes, there is an implementation, but it is work in progress and not public (yet). Feel free to contact me if you are interested.

Thanks again for your review.
