+++ docs/1_usage.rst
@@ -1,3 +1,4 @@
+.. _usage:
 
 =====
 Usage
