### 2020-06-04 e494354
include badges in documentation
