### 2020-06-09 4a20a0b
clarify canonical representation in paper.md
