### 2020-06-09 #2224 @MrMinimal64
I added some clarifications and examples of canonical form polynomial representations.
