### 2020-06-11 93274b6
Streamline array conversions, checks and dtypes
* Use asarray() instead of array() to avoid unnecessary copies
* Use pre-defined dtypes instead of np.<dtype> directly
