+++ tests/test_helpers.py
@@ -5,7 +5,7 @@ import sys
 import numpy as np
 
 from multivar_horner import HornerMultivarPolynomial, MultivarPolynomial
-from multivar_horner.global_settings import UINT_DTYPE, FLOAT_DTYPE
+from multivar_horner.global_settings import FLOAT_DTYPE, UINT_DTYPE
 from tests.test_settings import MAX_COEFF_MAGNITUDE, MAX_NUMERICAL_ERROR, NR_COEFF_CHANGES, NR_TEST_POLYNOMIALS
 
 
