+++ tests/test_it.py
@@ -24,7 +24,8 @@ from multivar_horner.multivar_horner import HornerMultivarPolynomial, MultivarPo
 # settings for numerical stability tests
 from tests.test_helpers import evaluate_numerical_error, naive_eval_reference, proto_test_case, vectorize
 from tests.test_settings import (
-    COEFF_CHANGE_DATA, DEGREE_RANGE, DIM_RANGE, INVALID_INPUT_DATA,
+    COEFF_CHANGE_DATA, DEGREE_RANGE, DIM_RANGE, INPUT_DATA_INVALID_TYPES_CONSTRUCTION,
+    INPUT_DATA_INVALID_TYPES_QUERY, INPUT_DATA_INVALID_VALUES_CONSTRUCTION, INPUT_DATA_INVALID_VALUES_QUERY,
     MAX_INP_MAGNITUDE, NR_TEST_POLYNOMIALS, TEST_RESULTS_PICKLE, VALID_TEST_DATA,
 )
 
@@ -95,19 +96,46 @@ class MainTest(unittest.TestCase):
     def test_invalid_input_detection(self):
 
         print('\n\nTEST INVALID INPUT DETECTION...')
-        for inp, expected_output in INVALID_INPUT_DATA:
-            coeff, exp, x = inp
-            with pytest.raises(AssertionError):
-                MultivarPolynomial(coeff, exp, rectify_input=False, validate_input=True, compute_representation=True)
 
-            with pytest.raises(AssertionError):
-                HornerMultivarPolynomial(coeff, exp, rectify_input=False, validate_input=True,
-                                         compute_representation=True, find_optimal=False)
+        def construction_should_raise(data, expected_error):
+            for inp, expected_output in data:
+                coeff, exp, x = inp
+                with pytest.raises(expected_error):
+                    MultivarPolynomial(coeff, exp, rectify_input=False, validate_input=True, )
+                with pytest.raises(expected_error):
+                    HornerMultivarPolynomial(coeff, exp, rectify_input=False, validate_input=True, find_optimal=False)
+                with pytest.raises(expected_error):
+                    HornerMultivarPolynomial(coeff, exp, rectify_input=False, validate_input=True, find_optimal=True)
 
-            with pytest.raises(AssertionError):
-                HornerMultivarPolynomial(coeff, exp, rectify_input=False, validate_input=True,
-                                         compute_representation=True, find_optimal=True)
+        construction_should_raise(INPUT_DATA_INVALID_TYPES_CONSTRUCTION, TypeError)
+        construction_should_raise(INPUT_DATA_INVALID_VALUES_CONSTRUCTION, ValueError)
 
+        # input rectification with negative exponents should raise a ValueError:
+        coeff = [3.2]
+        exp = [[0, 3, -4]]
+        with pytest.raises(ValueError):
+            HornerMultivarPolynomial(coeff, exp, rectify_input=True, validate_input=False, find_optimal=True)
+        with pytest.raises(ValueError):
+            MultivarPolynomial(coeff, exp, rectify_input=True, validate_input=False)
+        with pytest.raises(ValueError):
+            HornerMultivarPolynomial(coeff, exp, rectify_input=True, validate_input=False, find_optimal=False)
+
+        def query_should_raise(data, expected_error):
+            for inp, expected_output in data:
+                coeff, exp, x = inp
+                # construction must not raise an error:
+                p = MultivarPolynomial(coeff, exp, rectify_input=False, validate_input=True)
+                with pytest.raises(expected_error):
+                    p(x, rectify_input=False, validate_input=True)
+                p = HornerMultivarPolynomial(coeff, exp, rectify_input=False, validate_input=True)
+                with pytest.raises(expected_error):
[…]
