+++ multivar_horner/helpers_fcts_numba.py
@@ -11,6 +11,7 @@ import numpy as np
 from numba import b1, f8, jit, u4
 
 
+# TODO define and import dtypes globally
 @jit(f8(f8[:], f8[:], u4[:, :]), nopython=True, cache=True)
 def naive_eval(x, coefficients, exponents):
     nr_coeffs = len(coefficients)
