+++ tests/test_settings.py
@@ -1,6 +1,11 @@
 import os
 from math import log10
 
+import numpy as np
+from numpy import array as a
+
+from multivar_horner.global_settings import UINT_DTYPE
+
 EXPORT_RESOLUTION = 300  # dpi
 EXPORT_SIZE_X = 19.0  # inch
 EXPORT_SIZE_Y = 11.0  # inch
@@ -30,31 +35,120 @@ MAX_ERR_EXPONENT = max(-15, (int(log10(MAX_COEFF_MAGNITUDE)) - 10))
 MAX_NUMERICAL_ERROR = 10 ** MAX_ERR_EXPONENT
 
 # TEST CASES:
-INVALID_INPUT_DATA = [
-    # calling with x of another dimension
-    (([1.0, 2.0, 3.0],
-      [[3, 1, 0], [2, 0, 1], [1, 1, 1]],
-      [-2.0, 3.0]),
+
+# invalid:
+# rectify_input=False, numpy arrays needed
+# should raise TypeError
+INPUT_DATA_INVALID_TYPES_CONSTRUCTION = [
+    # not numpy.ndarray
+    ((([[1.0], [2.0], [3.0]]),
+      a([[3, 1, 0], [2, 0, 1], [1, 1, 1]], dtype=UINT_DTYPE),
+      a([-2.0, 3.0, 1.0])),
+     29.0),
+    ((a([[1.0], [2.0], [3.0]]),
+      ([[3, 1, 0], [2, 0, 1], [1, 1, 1]]),
+      a([-2.0, 3.0, 1.0])),
+     29.0),
+    ((None,
+      a([[3, 1, 0], [2, 0, 1], [1, 1, 1]], dtype=UINT_DTYPE),
+      a([-2.0, 3.0, 1.0])),
+     29.0),
+    ((a([[1.0], [2.0], [3.0]]),
+      None,
+      a([-2.0, 3.0, 1.0])),
      29.0),
 
-    (([1.0, 2.0, 3.0],
-      [[3, 1, 0], [2, 0, 1], [1, 1, 1]],
-      [-2.0, 3.0, 1.0, 4.0]),
+    # incorrect dtype
+    ((a([[1.0], [2.0], [3.0]], dtype=np.int),
+      a([[3, 1, 0], [2, 0, 1], [1, 1, 1]], dtype=UINT_DTYPE),
+      a([-2.0, 3.0, 1.0])),
      29.0),
-
-    # negative exponents are not allowed
-    (([1.0, 2.0, 3.0],
-      [[3, -1, 0], [2, 0, 1], [1, 1, 1]],
-      [-2.0, 3.0, 1.0, 4.0]),
+    ((a([[1.0], [2.0], [3.0]]),
+      a([[3, 1, 0], [2, 0, 1], [1, 1, 1]], dtype=np.float),
+      a([-2.0, 3.0, 1.0])),
      29.0),
-
-    # duplicate exponent entries are not allowed
-    (([1.0, 2.0, 3.0],
-      [[3, 1, 0], [3, 1, 0], [2, 0, 1], [1, 1, 1]],
-      [-2.0, 3.0, 1.0, 4.0]),
-     29.0),
-
 ]
+
+INPUT_DATA_INVALID_TYPES_QUERY = [
+    # not numpy.ndarray
+    ((a([[1.0], [2.0], [3.0]]),
+      a([[3, 1, 0], [2, 0, 1], [1, 1, 1]], dtype=UINT_DTYPE),
+      ([-2.0, 3.0, 1.0])),
+     29.0),
+    ((a([[1.0], [2.0], [3.0]]),
[…]
