### 2020-06-15 77c2a37
include badges in doc home
