### 2020-06-18 d13c95e
include doc badge
