+++ README.rst
@@ -7,6 +7,10 @@ multivar_horner
 .. image:: https://travis-ci.org/MrMinimal64/multivar_horner.svg?branch=master
     :target: https://travis-ci.org/MrMinimal64/multivar_horner
 
+.. image:: https://readthedocs.org/projects/multivar_horner/badge/?version=latest
+    :alt: documentation status
+    :target: https://multivar_horner.readthedocs.io/en/latest/?badge=latest
+
 .. image:: https://img.shields.io/pypi/wheel/multivar-horner.svg
     :target: https://pypi.python.org/pypi/multivar-horner
 
@@ -17,6 +21,7 @@ multivar_horner
 .. image:: https://img.shields.io/pypi/v/multivar_horner.svg
     :alt: latest version on PyPI
     :target: https://pypi.python.org/pypi/multivar-horner
+
 
 
 ``multivar_horner`` is a python package implementing a multivariate
