### 2020-06-23 #2224 @dpsanders
👋 @moorepants, @alex-konovalov and @henrik227, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Thanks!
