### 2020-06-25 #2392 @whedon
[REVIEW]: multivar_horner: A Python package for computing Horner factorisations of multivariate polynomials
**Submitting author:** @MrMinimal64 (<a href="http://orcid.org/0000-0002-1819-6975">Jannik Michelfeit</a>)
**Repository:** <a href="https://github.com/MrMinimal64/multivar_horner" target ="_blank">https://github.com/MrMinimal64/multivar_horner</a>
**Version:** 2.1.1

@henrik227, @alex-konovalov and @saschatimme, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:

---
