### 2020-06-29 6c7bf02
update code example in paper
v2.1.0 changed the function signature of .eval() to automatically convert the input into numpy arrays
