### 2020-06-29 ee35d2f
Update CHANGELOG.rst
fix bullet list
