### 2020-06-29 #2392 @MrMinimal64
@dpsanders please bump the sofware version to v2.1.0 (latest release). Some useful API changes/clarifications have been included: https://github.com/MrMinimal64/multivar_horner/blob/master/CHANGELOG.rst#210-2020-06-15
