+++ README.rst
@@ -3,8 +3,8 @@ multivar_horner
 ===============
 
 
-
 .. image:: https://travis-ci.org/MrMinimal64/multivar_horner.svg?branch=master
+    :alt: CI status
     :target: https://travis-ci.org/MrMinimal64/multivar_horner
 
 .. image:: https://readthedocs.org/projects/multivar_horner/badge/?version=latest
@@ -12,6 +12,7 @@ multivar_horner
     :target: https://multivar_horner.readthedocs.io/en/latest/?badge=latest
 
 .. image:: https://img.shields.io/pypi/wheel/multivar-horner.svg
+    :alt: wheel status
     :target: https://pypi.python.org/pypi/multivar-horner
 
 .. image:: https://pepy.tech/badge/multivar-horner
@@ -22,6 +23,9 @@ multivar_horner
     :alt: latest version on PyPI
     :target: https://pypi.python.org/pypi/multivar-horner
 
+.. image:: https://joss.theoj.org/papers/0b514c6894780f3cc81ed88c141631d4/status.svg
+    :alt: JOSS status
+    :target: https://joss.theoj.org/papers/0b514c6894780f3cc81ed88c141631d4
 
 
 ``multivar_horner`` is a python package implementing a multivariate
