### 2020-07-08 #2392 @henrik227
I am concerned that multivar_horner does not qualify as a `substantial scholarly effort' as described in the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort), for the following reasons:

- The project is not very large (see the statistics [here](https://github.com/openjournals/joss-reviews/issues/2224#issuecomment-632872231)).
- It seems to be a rather straight-forward implementation of an existing algorithm (described in Ceberio & Kreinovich, 2004).

However, I do believe that multivar_horner is useful for anyone who needs to evaluate large multivariate polynomials.

Initially, I thought that the search-based approach mention in the `further reading' section could be a major contribution of multivar_horner. However, according to the [documentation](https://multivar-horner.readthedocs.io/en/latest/3_optimal_factorisation.html#optimal-horner-factorisations), this approach does not seem to have any practical benefits (yet?).
