### 2020-07-10 184befb
clarify optimal factorisation results
fix citation
