### 2020-07-10 #2392 @MrMinimal64
hello Henrik,

I do respect your opinion, but I would like to briefly argue against it in the following.

> The project is not very large (see the statistics here).

How large a project needs to be in order to fulfil the JOSS review criteria is open for discussion.
What can be said definitively however, is that many projects with a similar or even smaller footprint are being considered:
https://github.com/openjournals/joss-reviews/issues/2428#issuecomment-653255875
https://github.com/openjournals/joss-reviews/issues/2438#issuecomment-653794412

> It seems to be a rather straight-forward implementation of an existing algorithm (described in Ceberio & Kreinovich, 2004).

First of all I would like to point out that ``multivar_horner`` contains many computational aspects not included in the solely theoretical outline of the algorithm in (Ceberio & Kreinovich, 2004).
The mathematical considerations only comprise the factorisation process.
How a found Horner factorisation can be represented and evaluated in a computationally efficient way is not within the scope of the literature.
Whether or not an implementation is “straight-forward”, in the end always remains highly subjective and even if it was, I think it could very well still qualify as “substantial scholarly effort”.

``multivar_horner`` can be considered as a substantial scholarly effort due to the following reasons:

- It is the first publicly available implementation of any kind of Horner factorisations for multivariate polynomials. Without it, the, as you call them, ``existing algorithms`` were existing actually only in theory or private and in absence of any benchmarks, tests and documentation.
- Because of their simple differentiation and integration, as well as their pleasant vector space structure, polynomials in multiple variables are a standard choice as interpolants and are fundamental in ordinary differential equation (ODE) and partial differential equation (PDE) solvers (see for example [1-4]).
- Horner factorisations have beneficial properties for handling polynomials as I am proving in the paper under review here
- How well all the theoretical approaches mentioned above can be “substantial” to academia depends on transferring them into practice by providing numerical availablility and applicability. That is exactly what this package does, which in turn makes it “substantial” to academia as well.
- That the package has been downloaded 11k times in total and on average 27 times a day in the last month (s. https://pepy.tech/project/multivar-horner) demonstrates that there is a demand for the project and its code quality is adequate

As an open source package helpful for a large part of the scientific community ``multivar_horner`` should hence be considered for publication in the JOSS.

[1] Floater, Michael S., and Kai Hormann. "Barycentric rational interpolation with no poles and high rates of approximation." Numerische Mathematik 107.2 (2007): 315-331.
[2] Cools, Ronald. "Constructing cubature formulae: the science behind the art." Acta numerica 6 (1997): 1-54.
[3] Temlyakov, Vladimir N. "Cubature formulas, discrepancy, and nonlinear approximation." Journal of Complexity 19.3 (2003): 352-391.
[4] Cools, Ronald. "Advances in multidimensional integration." Journal of computational and applied mathematics 149.1 (2002): 1-12.
