### 2020-07-10 #2392 @henrik227
To clarify: I do not question that the package is useful. However, I cannot say for sure that this package "represent[s] not less than three months of work for an individual", and from the review criteria it is not clear to me how the usefulness of the software is weighed against the effort that went into it. I am well aware that it is difficult to objectively quantify the development effort of software.

I do not have a strong opinion on this; I just want to make sure that the package meets the standards of JOSS. If @dpsanders and @alex-konovalov consider the package a "substantial scholarly effort", I am ok with that.
