### 2020-07-28 #2392 @MrMinimal64
in the meantime a preprint version has been published on [arXiv.org](http://arxiv.org/abs/2007.13152)
