### 2020-07-28 43ed235
mention ability to change coeffs in paper
