### 2020-07-29 385bda4
Update paper.md
add abstract
mention maple software offering multivariate Horner fuctionality
mention mathematica
