### 2020-07-29 #2392 @MrMinimal64
The current paper version now contains some additional aspects:
- an abstract also used to the arXiv paper version
- mentioning of ``Maple`` as a software already supporting multivariate Horner factorisations
- reference to ``Mathematica`` offering 1D Horner factorisations
- mention the ability to a posteriori change the coefficients of a factorised polynomial (the ``multivar_horner`` factorisation process is coefficient agnostic)

@saschatimme, @alex-konovalov,  what is the status of the review?
