### 2020-07-29 f680657
Update paper.md
mention Sascha Timme as reviewer
