### 2020-08-26 #2392 @henrik227
@dpsanders I was waiting on a decision regarding whether this project is considered a "substantial scholarly effort". Wouldn't it make sense to decide on that first?
