### 2020-08-26 #2392 @dpsanders
@henrik227 That's a fair point. In my opinion this submission **does** meet the criteria for a "substantial scholarly effort" and should be reviewed.
