### 2020-08-27 MrMinimal64/multivar_horner#17 @henrik227
JOSS submission: quality of writing
I found a few minor things in the paper that should be improved:

- The sentence "Representing multivariate polynomials of arbitrary degree also in canonical form, computing derivatives of polynomials and evaluating polynomials at a given point are further features of the package." doesn't read well; the problem is the "also". I would rephrase the entire sentence to something like "In addition, the package allows to/offers..."
- "can be helpful always when" should instead be "can be helpful whenever". In addition, perhaps I would change the "can be helpful" to "is useful" or so.
- "The polynomial p is the sum of 5 monomials" There are only four monomials.

This issue is part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/2392).
