### 2020-08-27 cc05dd7
paper: improve quality of writing
fix #17
