### 2020-08-27 MrMinimal64/multivar_horner#17 @MrMinimal64
Thanks for you feedback. Feel free to reopen if you have further suggestions.
