### 2020-08-27 MrMinimal64/multivar_horner#17 @saschatimme
Hi @MrMinimal64,

I also read through the paper for the JOSS review openjournals/joss-reviews#2392.

Here are some notes / comments / questions:

**p.1:**

> Polynomials are a central concept in mathematics and find application in a wide range of fields.

Maybe add some references?

**p.2:**
I think the way the number of operations are counted is not typical in the literature.
In particular, the evaluation of `x^n` is considered to be only a single operation.
The cost should rather the minimal number of multiplications necessary (roughly`ceil(log(n))` with power by squaring).

If the example on page 2 is computed with the `find_optimal` flag, then the result is
```
polynomial = HornerMultivarPolynomial(coefficients, exponents, find_optimal=True, compute_representation=True)
[#ops=10] p(x) = x_3 (x_1 (2.0 x_1 + 3.0 x_2)) + 5.0 + 1.0 x_1^3 x_2
```
which is actually a worse factorization than the example in the paper.

> This recipe encodes all operations required to evaluate the polynomial in numpy arrays (Walt, Colbert, & Varoquaux, 2011). This enables the use of functions just in time compiled by Numba (Lam, Pitrou, & Seibert, 2015), which cause the polynomial evaluation to be computationally efficient.

I was a little bit confused by how exactly the evaluation is done. My understanding of the source code is, that this is compiled to a bunch of arrays which encode the evaluation program and a for loop which iterates over these arrays and reads from the arrays, and performs either an addition or multiplication (and stores and reads input from some indices).
There is an alternative approach, where a straight line program is generated and compiled. It could be helpful to clarify that this is not done.

**p.4**

The possible number of terms of a multivariate polynomial in `n` variables of degree `m`  is `{n + m \choose n}`, not `(n + 1)^m`. A quadratic polynomial in 3 variables has 10 coefficients, not 27.
Is the `(n + 1)^m` count used in all the benchmarks / graphs?

I am confused by the experiment for the numerical evaluation error.
Are these absolute or relative errors?
If I put in the all-one vector in the naive evaluation formula of a multivariate polynomial, I expect to obtain as a result the sum of the coefficients. So, if there are 10^6 coefficients, I expect a relative accuracy of approx. `10^6 * 2^-53  \approx 10^-11` (assuming coefficients uniform drawn from `[-1,1]`) in line with the results for the naive evaluation.
How was the true sum of the coefficients determined? Just using 64 bit floating point arithmetic? Or a compensated sum scheme or extended precision?
I think a better experiment would be to use interval arithmetic for the evaluation and to compare the size of the intervals. Alternatively, evaluations at randomly chosen points would still be better than the all 1 vector. There, to assess the accuracy of the evaluation, the evaluation also needs to be performed with extended precision (128 bit should be sufficient).

(Peña and Sauer, 2000) show that their Horner scheme is forward and backward stable. Does this apply to the implemented method as well?
