### 2020-08-27 MrMinimal64/multivar_horner#17 @henrik227
Regarding the experiments with vectors that only contain 1: I was wondering too if that's a good experiment. I didn't think of it initially, but now I see the following problem: The function that computes `x^n` may check if `x` is 1 and simply return 1 in this case without computing anything. If that's the case, the potential loss of accuracy due to the computation of `x^n` wouldn't show up in the experiment at all.
While I'm not an expert in floating-point arithmetic, I suspect that even if `x^n` is always computed, the case where `x` is 1 is still a particularly favorable situation.
