### 2020-08-28 82230d5
improved quality of writing2
fix #17
added references to applications of polynomials
added remark about increasing the numerical stability to the documentation
used increased precision to compute ground truth of the numerical benchmarks
updated numerical benchmark plots
introduced separate numerical test file
