+++ paper/paper.md
@@ -37,7 +37,7 @@ Benchmarks additionally prove the advantages of the implementation and Horner fa
 
 # Introduction
 
-Polynomials are a central concept in mathematics and find application in a wide range of fields.
+Polynomials are a central concept in mathematics and find application in a wide range of fields[@prasolov2009polynomials; @boumova2002applications; @cools2002advances; @akritas1989elements; @Hecht1].
 (Multivariate) polynomials have different possible mathematical representations and the beneficial properties of some representations are in great demand in many applications[@LeeFactorization2013; @leiserson2010efficient; @Hecht1].
 
 The so called Horner factorisation is such a representation with beneficial properties.
@@ -68,9 +68,12 @@ The key functionality of `multivar_horner` is finding a good instance among the 
 Let's consider the example multivariate polynomial in canonical form $p(x) = 5 + 1 x_1^3 x_2^1 + 2 x_1^2 x_3^1 + 3 x_1^1 x_2^1 x_3^1$.
 The polynomial $p$ is the sum of $4$ monomials, has dimensionality $3$ and can also be written as $p(x) = 5 x_1^0 x_2^0 x_3^0 + 1 x_1^3 x_2^1 x_3^0 + 2 x_1^2 x_2^0 x_3^1 + 3 x_1^1 x_2^1 x_3^1$.
 The coefficients of the monomials are $5$, $1$, $2$ and $3$ respectively.
-It is trivial but computationally expensive to represent this kind of formulation with matrices and vectors and to evaluate it in this way.
-In this particular case for example a polynomial evaluation would require 27 operations.
-Due to its simplicity this kind of formulation is being used for defining multivariate polynomials as input.
+
+From this formulation it is straight forward to represent a multivariate polynomial with a single vector of coefficients and one exponent matrix.
+Due to its simplicity and universality this kind of representation is being used for defining polynomials as input.
+It should be noted that this most trivial representation is computationally very expensive to evaluate.
+In this particular case for example a polynomial evaluation based on arrays would in total require 27 operations (exponentiations, multiplications and additions).
+
 The following code snipped shows how to use `multivar_horner` for computing a Horner factorisation of $p$ and evaluating $p$ at a point $x$:
 
 ```python
@@ -98,9 +101,9 @@ The just in time compiled functions are always being used, since a pure python p
 
 
 It is important to note that in contrast to the one dimensional case, several concepts of degree exist for polynomials in multiple dimensions.
-Following the notation of [@trefethen2017multivariate] the usual notion of degree of a polynomial, the maximal degree, is the maximal sum of exponents of all monomials.
+Following the notation of [@trefethen2017multivariate] the usual notion of degree of a polynomial, the *total degree*, is the maximal sum of exponents of all monomials.
 This is equal to the maximal $l_1$-norm of all exponent vectors of the monomials.
-Accordingly the euclidean degree is the maximal $l_2$-norm and the maximal degree is the maximal $l_{\infty}$-norm of all exponent vectors.
+Accordingly the *euclidean degree* is the maximal $l_2$-norm and the *maximal degree* is the maximal $l_{\infty}$-norm of all exponent vectors.
 Refer to [@trefethen2017multivariate] for precise definitions.
 
 A polynomial is called fully occupied with respect to a certain degree if all possible monomials having a smaller or equal degree are present.
@@ -122,7 +125,7 @@ This effect intensifies as the dimensionality grows.
 
 For benchmarking our method the following procedure is used:
 In order to draw polynomials with uniformly random occupancy, the probability of monomials being present is picked randomly.
-For a fixed maximal degree $n$ in $m$ dimensions there are $(n+1)^m$ possible exponent vectors corresponding to monomials.
[…]
