### 2020-08-28 MrMinimal64/multivar_horner#17 @MrMinimal64
Thanks again for both of your feedback. I tried to fix as many of the issues you pointed out.
Please check out the latest paper version.

> evaluation of x^n is considered to be only a single operation

That is also what I am assuming. I realised that the formulation in the paper has been less than clear. I am referring to the total number of operations required to evaluate a polynomial based on arrays (including x^0, x^1...). I tried to clarify this in the paper. The number of operations of this most trivial representation should not be seen as a benchmark for the Horner factorisation in any way.

> [...] which is actually a worse factorization than the example in the paper.

In what way is the found “optimal” factorisation worse?

> There is an alternative approach, where a straight line program is generated and compiled

I guess this is what I was mimicking without knowing about this approach. Could you please point me in a direction where this is explained in detail? I guess that is what you have implemented in `StaticPolynomials`?!

> The possible number of terms of a multivariate polynomial in n variables of degree m is [...]]

This depends on the type of degree one is considering. You are right for the (typically used) total degree. In order to easily draw polynomials of a certain degree homogeneously, I am however refering to the maximal degree in all the experiments (cf. chapter "Degrees of multivariate polynomials" p.2).
Note: I mixed up the naming of the different degrees in the paper and fixed it.

> evaluation also needs to be performed with extended precision (128 bit should be sufficient).

Good point I am using 128 bit float now to compute the sum of coefficients and I updated the plots.

> evaluations at randomly chosen points would still be better than the all 1 vector

How can one know about the ground truth in this case?

> I suspect that […] the case where x is 1 is still a particularly favourable situation

It is true that this test case is not by any means ideal or without any bias.
However it is a test scenario in which the true function value and thereby also the numerical error is known.

> a better experiment would be to use interval arithmetic

I did not think of this and I believe this would be a good approach.
However the purpose of the benchmarks presented in the paper is solely to show the benefits of using Horner factorisations without making any specific performance claims.
In my opinion the current experiment setup fulfils this purpose. Do you agree?

> (Peña and Sauer, 2000) show that their Horner scheme is forward and backward stable. Does this apply to the implemented method as well?

It is not easy for me to assess for which set of Horner schemes these theoretical results hold and whether the implemented greedy scheme is included in this set.
In the publication of the greedy scheme theoretical considerations of numerical stability of the approach with interval arithmetic have been made (cf. Ceberio, M., & Kreinovich, V. (2004). Greedy algorithms for optimizing multivariate Hornerschemes).
