### 2020-08-28 #2392 @MrMinimal64
I tried to incorporate all the objections made by the reviewers in the latest version of the paper.
