### 2020-08-28 MrMinimal64/multivar_horner#17 @henrik227
> > evaluations at randomly chosen points would still be better than the all 1 vector
>
> How can one know about the ground truth in this case?

Similar to how you now compute the sum of coefficients with extended precision, as a reference you could evaluate the polynomials with extended precision.

Regarding how you count operations: Don't you think that counting the number of operations of exponentiation more precisely would improve your results? Intuitively, I would expect that in most cases the canonical form contains more exponentiations than the Horner factorization. If that's the case, the operation count for the canonical form should increase more than for the Horner factorization with a more precise count.
