### 2020-08-28 MrMinimal64/multivar_horner#17 @MrMinimal64
> evaluate the polynomials with extended precision

when I evaluate the numerical at random points uniformly sampled in [-1;1]^m I get these results:
![avg_numerical_error_VS_nr_coeff_rnd_samples](https://user-images.githubusercontent.com/15111305/91594806-c9354600-e962-11ea-9130-9f66183bfa50.png)
![numerical_error_heatmap_rnd_samples](https://user-images.githubusercontent.com/15111305/91594813-cb97a000-e962-11ea-96e1-8a54457c9a25.png)

From my assessment they are qualitatively similar, just with smaller numerical errors due to the smaller average sample magnitude and more noisy.
would you consider these experiments to be favorable?

> Regarding how you count operations [...]

It is not easy to reflect (and count) the actual amount of CPU operations performed for evaluating a mathematical expression, as this highly depends on many factors (datatypes, hardware, software in use, etc.).
cf. e.g. https://en.wikipedia.org/wiki/Exponentiation_by_squaring
