### 2020-08-31 MrMinimal64/multivar_horner#17 @henrik227
> would you consider these experiments to be favorable?

I have the impression that @saschatimme is more qualified than me to judge whether the results make sense, but I certainly prefer the evaluation at random points.

> It is not easy to reflect (and count) the actual amount of CPU operations performed for evaluating a mathematical expression, as this highly depends on many factors (datatypes, hardware, software in use, etc.).
> cf. e.g. https://en.wikipedia.org/wiki/Exponentiation_by_squaring

Sure, but I'm not suggesting that you try to count actual CPU operations. Since you are not generating straight-line code, I expect that there is quite a bit of overhead anyway. However, since this seems to be the convention for polynomials, I think it would make a lot more sense if you only counted additions and multiplications. It doesn't have to be exact, but at least it should reflect that computing `x^100` is more expensive than `x*y`.
