### 2020-08-31 MrMinimal64/multivar_horner#17 @saschatimme
Overall the changes look good! Here are some more comments:

> > It is not easy to reflect (and count) the actual amount of CPU operations performed for evaluating a mathematical expression, as this highly depends on many factors (datatypes, hardware, software in use, etc.).
> > cf. e.g. https://en.wikipedia.org/wiki/Exponentiation_by_squaring
>
> Sure, but I'm not suggesting that you try to count actual CPU operations. Since you are not generating straight-line code, I expect that there is quite a bit of overhead anyway. However, since this seems to be the convention for polynomials, I think it would make a lot more sense if you only counted additions and multiplications. It doesn't have to be exact, but at least it should reflect that computing `x^100` is more expensive than `x*y`.

I agree with @henrik227 here. The goal ist not to count CPU cycles (which is basically impossible), but rather have a solid (theoretical) model and to apply this. And here, the literature uses in my understanding the concept of counting the number of multiplications and additions necessary to evaluate a polynomial and this is also what I think should be used in the article and software as well.

> > [...] which is actually a worse factorization than the example in the paper.
>
> In what way is the found “optimal” factorisation worse?

This needs more operatations if you use the computational model described above.

As for the experiments, I am definitely more happy with using random points. One question I didn't see addressed: Are these errors relative or absolute?

> > There is an alternative approach, where a straight line program is generated and compiled
>
> I guess this is what I was mimicking without knowing about this approach. Could you please point me in a direction where this is explained in detail? I guess that is what you have implemented in `StaticPolynomials`?!

I unfortunately don't have a good reference for this, but this is basically how it was done in the good old fortran days in the 80s. You would use a computer algebra system (e.g. MAXIMA or REDUCE) to optimize your expression, then this would generate a fortran file containing the auto-generated source code to to evaluate your expression and then this got properly compiled and optimized. I [wrote](https://gist.github.com/saschatimme/436b02ee54121c85460b149ed71bfcc6) a little bit up how this kind of approach is also done in [HomotopyContinuation.jl](https://github.com/JuliaHomotopyContinuation/HomotopyContinuation.jl).
