### 2020-09-01 MrMinimal64/multivar_horner#17 @henrik227
> > There is an alternative approach, where a straight line program is generated and compiled
>
> I guess this is what I was mimicking without knowing about this approach. Could you please point me in a direction where this is explained in detail? I guess that is what you have implemented in `StaticPolynomials`?!

The way I understand it, you already represent the Horner factorization as a tree, where each node represents an operation. You should be able to generate straight line code from this tree by traversing it in post-order and generating an instruction/line of code at every node. Of course you need to make sure that the intermediate operands have unique names.
