### 2020-09-09 MrMinimal64/multivar_horner#17 @MrMinimal64
> the literature uses in my understanding the concept of counting the number of multiplications and additions necessary to evaluate a polynomial

for me this directly conflicts with ...

> It doesn't have to be exact, but at least it should reflect that computing x^100 is more expensive than x*y.

x^2 with 0 operations would then be considered to be favorable to x*x with 1 operation.
Due to the previously mentioned limitations of counting computational operations with any approach, the choice of a counting rule is not at all obvious. However to me it seems as if the rule currently applied, despite all its limitations, overall makes more sense.
In a way one could consider this to be an approximation of counting the amount of assembler or straight line code instructions required for an (idealized) evaluation program.

> One question I didn't see addressed ...

The errors are absolute. Do you think normalizing them makes more sense?

> I wrote a little bit up how this kind of approach [...]

Thanks. This is indeed a cool feature. Unfortunately I currently don't see how I could mimic straight line compilation in Python and I think this would even justify its own package.

> You should be able to generate straight line code from this tree by traversing it

That is exactly the implemented procedure, however ...

> generating an instruction/line of code at every node

... from my perspective cannot be done without hacky workarounds. Maybe I am not aware of some capabilities Python has.
