### 2020-09-09 1486f9c
improve quality of writing 2
address issues mentioned by the reviewers in #17
numerical error evaluation on a random point
ground truth generated with 128 precision
update plots
update benchmark description in the paper and in the documentation
mention errors being absolute
