+++ tox.ini
@@ -29,4 +29,4 @@ passenv =
 [testenv:codestyle]
 skip_install = true
 commands = multilint
-basepython = python3.7
+basepython = python3.8

[…]
