+++ .travis.yml
@@ -6,7 +6,7 @@ language: python
 matrix:
     include:
         # has to be the same py version specified in tox.ini
-        -   python: 3.7
+        -   python: 3.8
             env: TOXENV=codestyle
 
         -   python: 3.6
