### 2020-09-10 MrMinimal64/multivar_horner#17 @henrik227
> x^2 with 0 operations would then be considered to be favorable to x*x with 1 operation.

How do you get to 0 operations for `x^2`? Is it possible that you didn't use the base 2 logarithm? See https://en.wikipedia.org/wiki/Exponentiation_by_squaring#Computational_complexity.

> Thanks. This is indeed a cool feature. Unfortunately I currently don't see how I could mimic straight line compilation in Python and I think this would even justify its own package.
> [...]
> Maybe I am not aware of some capabilities Python has.

Ah, I see. I never tried it myself, but I think there are several options:

- Generating the function definition as a string, and then passing that string to [`eval`](https://docs.python.org/3.6/library/functions.html#eval), [`exec`](https://docs.python.org/3.6/library/functions.html#exec), or [`compile`](https://docs.python.org/3.6/library/functions.html#compile). I'm not too familiar with the differences between those functions, so I'm not sure which one is the best choice.
- Writing the function definition to a code file and then dynamically loading it at runtime with [`importlib.import_module`](https://docs.python.org/3.6/library/importlib.html#importlib.import_module). I could imagine that this might be tricky with the Python package infrastructure.

I don't know what is considered most "Pythonic", and if those methods are compatible with the JIT compilation you use. But perhaps the code is already good enough without JIT compilation. In addition, it would still be possible to simply write the code to a text file to be used in a different program. Given that the generated code should be relatively simple, one could even support other languages, for example C.
