### 2020-09-10 MrMinimal64/multivar_horner#17 @MrMinimal64
> you didn't use the base 2 logarithm

correct. As discussed, transforming code into hardware instructions is a complex topic (cf. "compiler construction", involves hardware capabilities etc.). Interpreters/compilers are complicated pieces of software designed for this exact purpose and it is best letting them decide how the operations are carried out internally.

> I think there are several options: [...]

These are the "hacky" workarounds I was referring to. As with the issue above, the problem is that one really has to put in a lot of work and care about tiny details to finally end up with more performant code than the original product and that on any given hardware.

Please do not get me wrong. I am not saying that the suggestions you made are impossible or without any advantages.
In my opinion however, the "overhead free" evaluation of mathematical expressions in Python would justify its own package or perhaps even multiple and this task is way beyond the scope of this package. In Theory

> simply write the code to a text file

 sounds quite easy. In reality I am sure, it is not that simple. Probably these are among the reasons why this is (to the best of my knowledge) not common practice .

Without being optimal, the approach implemented here (just in time compiled evaluation of straight line code encoded into numpy arrays) from my standpoint offers a good trade off between simplicity, flexibility and performance.
