### 2020-09-10 MrMinimal64/multivar_horner#17 @henrik227
> > you didn't use the base 2 logarithm
>
> correct. As discussed, transforming code into hardware instructions is a complex topic (cf. "compiler construction", involves hardware capabilities etc.). Interpreters/compilers are complicated pieces of software designed for this exact purpose and it is best letting them decide how the operations are carried out internally.

In general, what you say might be true, but I don't see how it is related to the question of which logarithm has to be used. As mention above, I think the operation count should reflect the theoretical complexity of the problem, and that's where the base 2 logarithm comes from. No matter what the compiler or hardware does, it seems very likely that `x^2` is computed one multiplication, and `x^8` with three (if `x` is a floating-point number).

> > I think there are several options: [...]

I'm not suggesting that you should implement any of this as part of the package. I just had the impression that you were interested in giving it a try.
I think it's an interesting project. Since Python already offers a `compile` function, this solution doesn't seem too "hacky" to me. I suspect that a "dynamic" evaluation of an expression the way you do it comes with a significant overhead in terms of control flow, so I would expect that it shouldn't be too difficult to get better performance by directly generating code.
Whether or not generating code as a text file is useful probably depends on the use case and the available alternatives. One would basically end with a compiler for a (very limited) domain-specific language, which is not unusual.
