### 2020-09-14 MrMinimal64/multivar_horner#17 @MrMinimal64
In my opinion, it would just be confusing for the reader to count the operations in a non intuitive way and a lengthy explanation of a counting method in the paper would unnecessarily overcomplicate things.
Since I am not trying to prove any theoretical lower bound for the number of operations, I guess the counting approach does not have to be very sophisticated. As a consequence I think the best choice is simply the counting procedure most intuitive for the reader.

> it's an interesting project [...]

I agree.

> One would basically end with a compiler for a (very limited) domain-specific language, which is not unusual.

Indeed. But this functionality ("compiling mathematical expressions in Python") is beyond the scope of this package.
