### 2020-09-14 d88103a
Update paper.md
fix #18
