### 2020-09-14 f5c87fb
Update paper.md
introduce line wrap for long code lines
