### 2020-09-14 e4d7b49
Update paper.md
update line wrap of string representation
