### 2020-09-14 4f13301
Update paper.md
update spacing of line wrap
