### 2020-09-17 MrMinimal64/multivar_horner#17 @henrik227
> In my opinion, it would just be confusing for the reader to count the operations in a non intuitive way and a lengthy explanation of a counting method in the paper would unnecessarily overcomplicate things.
> Since I am not trying to prove any theoretical lower bound for the number of operations, I guess the counting approach does not have to be very sophisticated. As a consequence I think the best choice is simply the counting procedure most intuitive for the reader.

Well, for me, using the logarithm is the more obvious and intuitive solution. I'm don't think it needs any explanation, and if you think it does, I'm sure it's possible to find a reference or explain it in a footnote.

Anyway, I don't think we will come to an agreement regarding this issue.
