### 2020-09-17 #2392 @henrik227
> 👋 @saschatimme and @henrik227: Many thanks for your very useful feedback in the issues on the repo! Are you happy with the current version?

I still disagree with how operations are counted in this package, but I think it's ok since it's explained both in the paper and the documentation.
