### 2020-09-17 MrMinimal64/multivar_horner#17 @henrik227
As a compromise, you could implement the method using the logarithm in addition to the existing one and allow the user to choose what they consider more meaningful.
