### 2020-09-17 MrMinimal64/multivar_horner#17 @saschatimme
> > One question I didn't see addressed ...
>
> The errors are absolute. Do you think normalizing them makes more sense?

Since floating point arithmetic only guarantees relative accuracy the relative error or normalized error makes more sense.
