### 2020-09-17 #2392 @saschatimme
I am in the same boat as @henrik227 that I disagree with the way operations are counted in the package. The article states
> multivar_horner implements a multivariate Horner scheme using the greedy heuristic presented in (Ceberio & Kreinovich, 2004).

However, looking into  (Ceberio & Kreinovich, 2004) the way operations are counted there is actually how @henrik227 and I suggest they are supposed to be counted (to compute `x^3` takes 2 operations) and not as the author implemented it and describes it in the article where `x^3` only requires single operation.
