### 2020-09-17 7bc77b4
count only multiplications
as discussed in #17:
count only multiplications
the amount of additions stays constant irrespective of the factorisation
instead of counting just one operation for every exponentiation, count them as exponent-1 multiplications
