### 2020-09-17 4483ee3
update the string representations
reflecting the changes in counting the amount of required operations
