### 2020-09-17 7592a9a
display relative errors
as discussed in #17
update the plots and description in the paper
