### 2020-09-18 cbd78ed
update coefficient amount comparison plot
update plots in order to reflect the changes in counting the amount of coefficients
#17
