### 2020-09-18 #2392 @MrMinimal64
The operations are now being counted as you suggest (solely the number of multiplications).
The optimal search heuristic now minimises the new operation count and finds "optimal" factorisations in this sense.
I also evaluate the relative numerical error in the benchmarks.
The paper as well as the documentation have been updated accordingly.
