### 2020-09-18 7d9f3ae
change figure positioning in paper
