### 2020-09-18 #2392 @MrMinimal64
I updated the placement of the figures (now the paper has a length of 8 pages again)
