### 2020-09-18 MrMinimal64/multivar_horner#17 @MrMinimal64
I have incorporated the changes you suggested.
Thanks for your honest and helpful feedback!
All the best for your work.
