### 2020-09-27 #2392 @dpsanders
I don't see a discussion in the paper about how the number of operations is counted.

@saschatimme and @henrik227, are you happy with the current version?

@alex-konovalov: Are you able to complete your review? If not, I will remove you as a reviewer.
