### 2020-09-27 #2392 @MrMinimal64
> I don't see a discussion in the paper about how the number of operations is counted.

it started here:
https://github.com/MrMinimal64/multivar_horner/issues/17#issuecomment-681935918
Now the method of counting the operations is as the reviewers have suggested.
