### 2020-09-27 #2392 @dpsanders
Yes thanks; I would like the paper to describe the implemented choice for counting operations. (Maybe it's already there but I didn't see it.)
