### 2020-09-28 #2392 @henrik227
> Yes thanks; I would like the paper to describe the implemented choice for counting operations. (Maybe it's already there but I didn't see it.)

It's in the middle of page 2.

> @henrik227, are you happy with the current version?

I have some minor comments:

- Is it intentional that in the example right above the description of how operations are counted, the "old" way of counting operations is used? I'm referring to "27 operations (exponentiations, multiplications and additions)"
- "the number of additions [...] remains constant irrespective of the polynomial factorisation." To me, it's not obvious why this is true. Would it be possible to give some intuition or provide a reference?
- "The number of multiplications contained in the canonical representation of a polynomial consequently is equal to the sum of its exponents." Same as above. However, I don't think it's very important, so perhaps I would just remove it.
