### 2020-09-28 6fbe60e
update operation count explanation
in regard to #17
