### 2020-09-28 #2392 @MrMinimal64
> I have some minor comments:

I tried to address all of your remarks. Please check if your concerns persist.
