### 2020-09-30 #2392 @MrMinimal64
Thanks for the fair review!

@dpsanders are there any open points?
