### 2020-09-30 #2392 @dpsanders
It looks good to me!

Could you please create a software archive, e.g. at zenodo.org, and paste the version number and the DOI here? Thanks!
