+++ paper/paper.bib
@@ -52,6 +52,7 @@ series = {LLVM ’15}
   number={109},
   pages={308--335},
   year={1819},
+   doi = {10.1098/rstl.1819.0023},
   publisher={The Royal Society London}
 }
 
@@ -104,6 +105,7 @@ series = {LLVM ’15}
 	Title = {Fast Interpolation and {F}ourier Transform in High-Dimensional Spaces},
 	Volume = {857},
 	Year = {2018},
+	 doi = {10.1007/978-3-030-01177-2_5},
 }
 
 @article{trefethen2017multivariate,
@@ -121,6 +123,7 @@ series = {LLVM ’15}
   title     = {Greedy algorithms for optimizing multivariate {H}orner schemes},
   year      = {2004},
   publisher = {Association for Computing Machinery (ACM)},
+  doi={10.1145/980175.980179},
 }
 
 
@@ -151,6 +154,7 @@ the polynomial basis is chosen suitably.},
   pages     = {313--322},
   abstract  = {As an extension of our previous paper [10] which gave for-ward and backward error estimates, we perform a running error analysis ofthe multivariate Horner scheme.  This leads to a modified algorithm whichcomputes the value of the polynomial together with an error estimate.},
   publisher = {Springer},
+  doi={10.1007/s006070070002},
 }
 
 
@@ -163,6 +167,7 @@ the polynomial basis is chosen suitably.},
   pages={29--54},
   year={2008},
   publisher={The Operations Research Society of Japan},
+  doi={10.15807/jorsj.51.29},
 }
 
 
@@ -180,6 +185,7 @@ polynomials and their derivatives is obtained. The schemes of computation are
 represented by trees because this type of graph describes exactly in which order
 the computations must be done. Some examples of algorithms for one and two
 variables are given.},
+doi={10.1090/s0025-5718-1990-0993925-1},
 }
 
 
@@ -200,6 +206,7 @@ better algorithms has proved difficult. In trying to improve upon the greedy sch
 plemented Monte Carlo tree search, a recent search method from the field of artificial intelligence.
 This results in better Horner schemes and reduces the cost of evaluating polynomials, sometimes by
 factors up to two.},
+doi={10.1016/j.cpc.2013.05.008},
 }
 
 @PhdThesis{LeeFactorization2013,
@@ -217,6 +224,7 @@ factors up to two.},
   pages={342--353},
   year={2010},
   organization={Springer}
+  doi={10.1007/978-3-642-15582-6_55}
 }
 
 # applications polynomials
@@ -245,6 +253,7 @@ factors up to two.},
   pages={1--12},
   year={2002},
   publisher={Elsevier}
+  doi={10.1016/s0377-0427(02)00517-4},
 }
 
 
