+++ paper/paper.bib
@@ -223,8 +223,8 @@ doi={10.1016/j.cpc.2013.05.008},
   booktitle={International Congress on Mathematical Software},
   pages={342--353},
   year={2010},
-  organization={Springer}
-  doi={10.1007/978-3-642-15582-6_55}
+  organization={Springer},
+  doi={10.1007/978-3-642-15582-6_55},
 }
 
 # applications polynomials
@@ -252,7 +252,7 @@ doi={10.1016/j.cpc.2013.05.008},
   number={1},
   pages={1--12},
   year={2002},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/s0377-0427(02)00517-4},
 }
 
@@ -261,6 +261,6 @@ doi={10.1016/j.cpc.2013.05.008},
   title={Elements of computer algebra with applications},
   author={Akritas, Alkiviadis},
   year={1989},
-  publisher={John Wiley \& Sons, Inc.}
+  publisher={John Wiley \& Sons, Inc.},
 }
 
