### 2020-09-30 #2392 @MrMinimal64
The last remaining DOI is actually incorrect. Many books have this name.
