### 2020-09-30 #2392 @dpsanders
It seems to be https://doi.org/10.1017/CBO9780511612916 ?
