+++ paper/paper.bib
@@ -71,7 +71,8 @@ series = {LLVM ’15}
   title={Introduction to numerical analysis},
   author={Neumaier, Arnold},
   year={2001},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={10.1017/CBO9780511612916},
 }
 
 
