### 2020-09-30 #2392 @MrMinimal64
The current version number is v2.1.1
and the DOI 10.5281/zenodo.4060365
