+++ docs/index.rst
@@ -26,6 +26,9 @@ multivar_horner
     :alt: JOSS status
     :target: https://joss.theoj.org/papers/0b514c6894780f3cc81ed88c141631d4
 
+.. image:: https://zenodo.org/badge/155578190.svg
+   :target: https://zenodo.org/badge/latestdoi/155578190
+
 
 a python package implementing a multivariate `Horner scheme ("Horner's method", "Horner's rule") <https://en.wikipedia.org/wiki/Horner%27s_method>`__:cite:`horner1819xxi` for efficiently evaluating multivariate polynomials.
 
