### 2020-10-01 #2392 @Kevin-Mattheus-Moerman
@MrMinimal64 I'm processing your paper for acceptance in JOSS. I have read your paper and ask you to work on the remaining minor points below:

- [ ] Make edit the ZENODO archive meta-data and make sure that in terms of title and author list it matches that of the paper

- [ ] The version tag for the archived version should also match the reviewed software version tag. Should we update the version tag of the reviewed software to `v2.1.1-zenodo` (we can do this for you) or should the ZENODO archive version tag be changed to `v2.1.1`?

- [ ] Please add city and country information to your affiliations

- [ ] Your paper is about to be processed for acceptance. I suggest you also proofread the draft one more time before we move on, and make any minor changes needed. In particular make sure the authors, affiliations, and acknowledgements are accurate/complete.

Kevin Moerman
_AEiC JOSS_
