### 2020-10-01 e9e370b
add city and country information to affiliations
