### 2020-10-01 0670a03
mark own reference as thesis
