+++ paper/paper.bib
@@ -76,12 +76,13 @@ series = {LLVM ’15}
 }
 
 
-@article{michelfeitresponse,
+@thesis{michelfeitresponse,
   title={A Response Surface Method based on Multivariate Newton Interpolation},
   author={Michelfeit, Jannik},
   year={2019},
   url = {https://mosaic.mpi-cbg.de/docs/Michelfeit2019.pdf},
   urldate = {2020-05-28},
+  type={Student Thesis},
 }
 
 @article{Hecht1,
