### 2020-10-01 #2392 @MrMinimal64
I uploaded the version `v2.1.1` to zenodo. The difference is just one missing DOI in a reference, but this should not make a difference for the software archive right? I think this version tag is preferable.
The DOI now changed to:
10.5281/zenodo.4062422
I updated the metadata of the software archive and added city and country to my affiliations.
The authors, affiliations, and acknowledgements should now be complete.
