### 2020-10-02 #2392 @MrMinimal64
Thank you all for enabling my first sole publication!

One thing I noticed. Since there is an underscore in the title, using the linked BibTex silently breaks a latex document.
I think it is very important to fix this, since the readers otherwise search for errors in their documents forever.
All underscores need to be escaped:

`title = {multivar\_horner: A Python package for computing Horner factorisations of multivariate polynomials},`
